requirements.txt
Django>=2.0
psycopg2
djangorestframework
djangorestframework-jwt
django-rest-auth
django-sendgrid-v5
drf-nested-routers
